[Edit] Change Map; Add alt map as Prefab; Refactor Map Handling
