+++ DungeonDuell/Assets/DungeonDuell/Prefab/CardsPhase/Maps/MapPackOld.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 771e9da516a153a429e2c9b412d49ea8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
