+++ DungeonDuell/Assets/DungeonDuell/Prefab/CardsPhase/Maps/MapPackVert.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: af14a35051cee7241acc6f7ce37547d8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
