+++ DungeonDuell/Assets/DungeonDuell/Prefab/DungeonPhase/Room/VirtualCameraP1.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 578585bde1200c141924bc722d647738
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
