+++ DungeonDuell/Assets/DungeonDuell/Script/CardPhase/TileClickHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Cinemachine;
+using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.Serialization;
@@ -154,6 +155,12 @@ namespace dungeonduell
 
         public bool SpawnTile(Vector3 mouseWorldPos, Card card, bool playerMove, bool spawnSourroundSetables, int owner)
         {
+            if (_tilemap == null)
+            {
+                _tilemap = FindFirstObjectByType<Grid>().GetComponentsInChildren<Tilemap>()
+                    .FirstOrDefault(tm => tm.gameObject.CompareTag(tileMapTag));
+            }
+
             var cellPosition =
                 _tilemap.WorldToCell(new Vector3(mouseWorldPos.x, mouseWorldPos.y, cam.transform.position.z));
             var clickedTile = _tilemap.GetTile(cellPosition);
@@ -260,10 +267,10 @@ namespace dungeonduell
 
                 if (spawnSourroundSetables)
                 {
-                    if(!hitContested) 
+                    if (!hitContested)
                         SpawnShadowTileOperations(playerMove, cellPosition, clickedTile);
 
-                    SpawnSetAbleOperation(cellPosition, clickedTile, owner,hitContested);
+                    SpawnSetAbleOperation(cellPosition, clickedTile, owner, hitContested);
 
                     if (playerMove)
                     {
@@ -300,15 +307,15 @@ namespace dungeonduell
             return false;
         }
 
-        private void SpawnSetAbleOperation(Vector3Int cellPosition, TileBase clickedTile, int owner,bool contestedHit)
+        private void SpawnSetAbleOperation(Vector3Int cellPosition, TileBase clickedTile, int owner, bool contestedHit)
         {
             foreach (var sourrendTilePos in GetSouroundCorr(cellPosition, currentDoorDir))
             {
                 var souroundTile = _tilemap.GetTile(sourrendTilePos.Item1);
-                
+
                 if (souroundTile == resetTile && contestedHit)
                 {
-                    _tilemap.SetTile(sourrendTilePos.Item1, setAbleTiles[owner -1]);
+                    _tilemap.SetTile(sourrendTilePos.Item1, setAbleTiles[owner - 1]);
                 }
                 else if (souroundTile == resetTile || shadowSetAbleTiles.Contains(souroundTile))
                 {
@@ -362,7 +369,6 @@ namespace dungeonduell
                         else if (souroundTile != shadowSetAbleTiles[i])
                             _tilemap.SetTile(sourrendTilePos.Item1, shadowSetAbleTiles[^1]);
                     }
-                       
                 }
 
                 if (souroundTile == resetTile && playerMove)
@@ -406,7 +412,7 @@ namespace dungeonduell
             {
                 if (tileBaseType == setAbleTiles[i])
                 {
-                    if (i < setAbleTiles.Length -1)
+                    if (i < setAbleTiles.Length - 1)
                     {
                         setAbleCount[i]++;
                     }
