+++ DungeonDuell/Assets/DungeonDuell/Script/CrossSystems/ConnectionsCollector.cs
@@ -19,28 +19,6 @@ namespace dungeonduell
             DontDestroyOnLoad(gameObject);
         }
 
-        private void OnEnable()
-        {
-            //Debug.Log("OnEnable called");
-            SceneManager.sceneLoaded += OnSceneLoaded;
-        }
-
-
-        private void OnDisable()
-        {
-            SceneManager.sceneLoaded -= OnSceneLoaded;
-        }
-
-        // called second
-        private void OnSceneLoaded(Scene scene, LoadSceneMode mode)
-        {
-            // So the map is not seen in Dunegeon Phase
-            if (scene.buildIndex == 2)
-                transform.GetChild(0).gameObject.SetActive(false);
-            else
-                transform.GetChild(0).gameObject.SetActive(true);
-        }
-
         public void AddRoom(Vector3Int pos, List<RoomConnection> conncection, RoomType type, RoomElement element,
             List<ConnectionDir> newAllowedDoors, int owner)
         {
