+++ DungeonDuell/UserSettings/EditorUserSettings.asset
@@ -27,10 +27,10 @@ EditorUserSettings:
       value: 005302575c510f0b5b5a092711730d44174e1b7e7f2a23367d284c67bae1603a
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 06040703550c0b09590d597013755944464f1e7d282a71367c7c1f63e0e66c61
+      value: 5b08525000045c080f5a0d7a1327094415154a7d7c7e2464792a1c37b7b53761
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 5b08525000045c080f5a0d7a1327094415154a7d7c7e2464792a1c37b7b53761
+      value: 06040703550c0b09590d597013755944464f1e7d282a71367c7c1f63e0e66c61
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 5757570006515f5e5f5e5d7b43735e4442151d2e782a24627d7b4e35e1e5646a
